implement merging of a mesh into another.
For data fields without boundary conditions, this only concatenates the element lists and updates the number of elements.
If there are boundary conditions present it will check if faces with non-'E' BCs can be matched together and connected with a 'E' BC.
It does not work yet with 'P' conditions yet as this will require some extra code.